managing user sessions

+++ ch4_processing_http_requests/javascript/main.js
@@ -1,3 +1,7 @@
+if (localStorage.getItem("user-token") == null) {
+	window.location.replace(document.location.origin + "/login/");
+}
+
 function renderItems(items, processType, elementId, processFunction) {
 	let itemsMeta = [];
 	let placeholder = "<div>";
@@ -35,29 +39,33 @@ function apiCall(url, method) {
 	xhr.withCredentials = true;
 	xhr.addEventListener("readystatechange", function () {
 		if (this.readyState === this.DONE) {
-			renderItems(
-				JSON.parse(this.responseText)["pending_items"],
-				"edit",
-				"pendingItems",
-				editItem
-			);
-			renderItems(
-				JSON.parse(this.responseText)["done_items"],
-				"delete",
-				"doneItems",
-				deleteItem
-			);
-			document.getElementById("completeNum").innerHTML = JSON.parse(
-				this.responseText
-			)["done_item_count"];
-			document.getElementById("pendingNum").innerHTML = JSON.parse(
-				this.responseText
-			)["pending_item_count"];
+			if (this.status === 401) {
+				window.location.replace(document.location.origin + "/login/");
+			} else {
+				renderItems(
+					JSON.parse(this.responseText)["pending_items"],
+					"edit",
+					"pendingItems",
+					editItem
+				);
+				renderItems(
+					JSON.parse(this.responseText)["done_items"],
+					"delete",
+					"doneItems",
+					deleteItem
+				);
+				document.getElementById("completeNum").innerHTML = JSON.parse(
+					this.responseText
+				)["done_item_count"];
+				document.getElementById("pendingNum").innerHTML = JSON.parse(
+					this.responseText
+				)["pending_item_count"];
+			}
 		}
 	});
 	xhr.open(method, url);
 	xhr.setRequestHeader("content-type", "application/json");
-	xhr.setRequestHeader("user-token", "token");
+	xhr.setRequestHeader("user-token", localStorage.getItem("user-token"));
 
 	return xhr;
 }
